Still fixing stuff.

+++ test/hilbert.jl
@@ -1,5 +1,5 @@
 function hilbert(n::Int)
-    A = zeros(Rational{BigInt},n,n)
+    A = zeros(Rational{Int},n,n)
     for i=1:n
         for j=1:n
             A[i,j] = 1 // (i+j-1)
